refactor modules + set up graph lib
